Update stepper color utils, update comments
